% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amce.R
\name{amce}
\alias{amce}
\title{Tidy estimation of AMCEs}
\usage{
amce(data, formula, id = NULL, weights = NULL, feature_order = NULL,
  feature_labels = NULL, level = 0.95, ...)
}
\arguments{
\item{data}{A data frame containing variables specified in \code{formula}. All RHS variables should be factors; the base level for each will be used in estimation and its reported AMCE will be zero (for printing).}

\item{formula}{A formula specifying an AMCE model to be estimated. All variables should be factors.}

\item{id}{An RHS formula specifying a variable holding respondent identifiers, to be used for clustering standard errors.}

\item{weights}{An (optional) RHS formula specifying a variable holding survey weights.}

\item{feature_order}{An (optional) character vector specifying the names of feature (RHS) variables in the order they should be encoded in the resulting data frame.}

\item{feature_labels}{A named list of \dQuote{fancy} feature labels to be used in output. By default, the function looks for a \dQuote{label} attribute on each variable in \code{formula} and uses that for pretty printing. This argument overrides those attributes or otherwise provides fancy labels for this purpose. This should be a list with names equal to variables in \code{formula} and character string values; arguments passed here override variable attributes.}

\item{level}{A numeric value indicating the significance level at which to calculate confidence intervals for the AMCEs (by default 0.95, meaning 95-percent CIs are returned).}

\item{\dots}{Additional arguments to \code{\link[stats]{glm}} or \code{\link[survey]{svyglm}}, the latter being used if \code{weights} is non-NULL.}
}
\value{
A data frame
}
\description{
Estimate AMCEs for a conjoint analysis and return a tidy data frame of results
}
\details{
Users may desire to specify a \code{family} argument via \code{\dots}, which should be a \dQuote{family} object such as \code{\link[stats]{gaussian}}. Sensible alternatives are \code{\link[stats]{binomial}} (for binary outcomes) and \code{\link[stats]{quasibinomial}} (for weighted survey data).
}
\examples{
data(hainmueller)
set.seed(12345)
cj(hainmueller, ChosenImmigrant ~ Gender + Education + LanguageSkills, 
   id = ~ CaseID, feature_order = c("LanguageSkills", "Gender", "Education"))
 
}
\seealso{
\code{\link{plot.cj_amce}}
}
