% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-cregg.R
\docType{package}
\name{cregg-package}
\alias{cregg-package}
\alias{cregg}
\alias{cj}
\title{Simple Conjoint Analyses and Visualization}
\usage{
cj(data, formula, id = NULL, weights = NULL, estimate = c("amce", "mm"),
  feature_labels = NULL, level = 0.95, ...)
}
\arguments{
\item{data}{A data frame containing variables specified in \code{formula}. All RHS variables should be factors; the base level for each will be used in estimation and for AMCEs the base level's AMCE will be zero.}

\item{formula}{A formula specifying a model to be estimated. All variables should be factors.}

\item{id}{An RHS formula specifying a variable holding respondent identifiers, to be used for clustering standard errors.}

\item{weights}{An (optional) RHS formula specifying a variable holding survey weights.}

\item{estimate}{A character string specifying an estimate type. Current options are average marginal component effects (or AMCEs, \dQuote{amce}, estimated via \code{\link{amce}}) or marginal means (or AMMs, \dQuote{mm}, estimated via \code{\link{mm}}). Additional options may be made available in the future.}

\item{feature_labels}{A named list of \dQuote{fancy} feature labels to be used in output. By default, the function looks for a \dQuote{label} attribute on each variable in \code{formula} and uses that for pretty printing. This argument overrides those attributes or otherwise provides fancy labels for this purpose. This should be a list with names equal to variables in \code{formula} and character string values; arguments passed here override variable attributes.}

\item{level}{A numeric value indicating the significance level at which to calculate confidence intervals for the AMCEs (by default 0.95, meaning 95-percent CIs are returned).}

\item{\dots}{Additional arguments to \code{\link{amce}} or \code{\link{mm}}.}
}
\description{
Simple analyses of conjoint (factorial) experiments and visualization of results.
}
\details{
The main function \code{cj} is a convenience function wrapper around the underlying estimation functions that provide for average marginal component effects (AMCEs), by default, via the \code{\link{amce}} function and marginal means (MMs) via the \code{\link{mm}} function. Additional estimands may be supported in the future through their own functions and through the \code{cj} interface. Plotting is provided via ggplot2 for both types of estimates.
}
\examples{
\dontrun{
# load data
requireNamespace("ggplot2")
data("hainmueller")

# calculate MMs
d1 <- cj(hainmueller, ChosenImmigrant ~ Gender + Education + 
         LanguageSkills + CountryOfOrigin + Job + JobExperience + 
         JobPlans + ReasonForApplication + PriorEntry, id = ~ CaseID,
         estimate = "mm")

# plot MMs
plot(d1)

# MMs split by profile number
stacked <- do.call("rbind", lapply(1:3, function(x) {
    out <- mm(hainmueller[hainmueller$contest_no == x,], ChosenImmigrant ~ Gender + 
              Education + LanguageSkills + CountryOfOrigin + Job + JobExperience + 
              JobPlans + ReasonForApplication + PriorEntry, id = ~ CaseID)
    out$contest <- x
    out
}))

## plot with grouping
plot(stacked, group = "contest", feature_headers = FALSE)

## plot with facetting
plot(stacked) + ggplot2::facet_wrap(~contest, nrow = 1L)

# estimate AMCEs
d2 <- cj(hainmueller, ChosenImmigrant ~ Gender + Education + 
         LanguageSkills + CountryOfOrigin + Job + JobExperience + 
         JobPlans + ReasonForApplication + PriorEntry, id = ~ CaseID)

# plot AMCEs
plot(d2)
}
}
\seealso{
\code{\link{amce}} \code{\link{mm}} \code{\link{plot.amce}}
}
\author{
Thomas J. Leeper <thosjleeper@gmail.com>
}
\keyword{package}
