% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffs.R
\name{amce_diffs}
\alias{amce_diffs}
\alias{cj_anova}
\title{Preference Heterogeneity Diagnostics}
\usage{
amce_diffs(data, formula, by, id = NULL, weights = NULL,
  feature_order = NULL, feature_labels = NULL,
  level_order = c("ascending", "descending"), alpha = 0.05, ...)

cj_anova(data, formula, id = NULL, weights = NULL, by = NULL, ...)
}
\arguments{
\item{data}{A data frame containing variables specified in \code{formula}. All RHS variables should be factors; the base level for each will be used in estimation and for AMCEs the base level's AMCE will be zero.}

\item{formula}{A formula specifying a model to be estimated. All variables should be factors.}

\item{by}{A formula containing only RHS variables, specifying grouping factors over which to perform estimation.}

\item{id}{Ignored. An RHS formula specifying a variable holding respondent identifiers, to be used for clustering standard errors.}

\item{weights}{An (optional) RHS formula specifying a variable holding survey weights.}

\item{\dots}{Additional arguments to \code{\link{amce}}, \code{\link{freqs}}, or \code{\link{mm}}.}

\item{estimate}{A character string specifying an estimate type. Current options are average marginal component effects (or AMCEs, \dQuote{amce}, estimated via \code{\link{amce}}), display frequencies (\dQuote{freq}, estimated via \code{\link{freqs}}), or marginal means (or AMMs, \dQuote{mm}, estimated via \code{\link{mm}}). Additional options may be made available in the future.}
}
\value{
\code{amce_diffs} returns a data frame similar to the one returned by \code{\link{cj}}, including a \code{BY} column for easy merging with results returned by that function. \code{cj_anova} returns an \code{\link[stats]{anova}} object.
}
\description{
Tests for preference heterogeneity in conjoint experiments
}
\details{
\code{cj_anova} takes a model formula (\dQuote{reduced} model) and generates a \dQuote{full} model with full saturated interactions between the variables specified in \code{by} and all RHS variables in \code{formula}, then computes an F-test comparing the two models, providing a test for whether preferences vary across levels of \code{by}. This is, in essence, a test of whether all such interaction coefficients are distinguishable from zero. Because the test depends on overall model fit, not the coefficient variances, clustering is irrelevant.
}
\examples{
# Test for heterogeneity by profile order
cj_anova(hainmueller, ChosenImmigrant ~ Gender + Education + LanguageSkills, by = ~ contest_no)

# Test for heterogeneity by CountryOfOrigin feature
cj_anova(hainmueller, ChosenImmigrant ~ Gender + Education, by = ~ CountryOfOrigin)

# Differences in AMCEs
amce_diffs(hainmueller, ChosenImmigrant ~ LanguageSkills + Education, ~ Gender, id = ~ CaseID)

}
\seealso{
\code{\link{amce}} \code{\link{mm}} \code{\link{freqs}} \code{\link{plot.cj_amce}}
}
\author{
Thomas J. Leeper <thosjleeper@gmail.com>
}
