% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq.R
\name{freqs}
\alias{freqs}
\title{Conjoint feature frequencies}
\usage{
freqs(data, formula, id, weights = NULL, feature_order = NULL,
  feature_labels = NULL, ...)
}
\arguments{
\item{data}{A data frame containing variables specified in \code{formula}. All RHS variables should be factors.}

\item{formula}{An RHS formula specifying conjoint features to tabulate. All RHS variables should be factors.}

\item{id}{An RHS formula specifying a variable holding respondent identifiers, to be used for clustering standard errors.}

\item{weights}{An (optional) RHS formula specifying a variable holding survey weights.}

\item{feature_order}{An (optional) character vector specifying the names of feature (RHS) variables in the order they should be encoded in the resulting data frame.}

\item{feature_labels}{A named list of \dQuote{fancy} feature labels to be used in output. By default, the function looks for a \dQuote{label} attribute on each variable in \code{formula} and uses that for pretty printing. This argument overrides those attributes or otherwise provides fancy labels for this purpose. This should be a list with names equal to variables in \code{formula} and character string values; arguments passed here override variable attributes.}

\item{\dots}{Ignored.}
}
\description{
Tabulate and visualize conjoint feature frequencies
}
\details{
\code{freqs} provides a descriptive check on the presentation of conjoint features (to ensure equal or intentionally unequal appearance of levels). This is mostly useful for plotting functionality provided in \code{\link{plot.cj_freqs}}.
}
\examples{
data(hainmueller)
freqs(hainmueller, ~ Gender + Education + LanguageSkills, id = ~ CaseID)
}
\seealso{
\code{\link{plot.cj_mm}}
}
