% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.amce}
\alias{plot.amce}
\alias{plot.mm}
\title{Plot AMCE estimates and MM descriptives}
\usage{
\method{plot}{amce}(x, group = NULL, feature_headers = TRUE,
  header_fmt = "(\%s)", size = 1, xlab = "Estimated AMCE", ylab = "",
  legend_title = "Feature", xlim = NULL, vline = 0,
  vline_color = "gray", theme = ggplot2::theme_bw(), ...)
}
\arguments{
\item{x}{A data frame returned from \code{\link{cj}} or \code{\link{mm}}.}

\item{group}{Optionally a character string specifying a grouping factor. This is useful when, for example, subgroup analyses or comparing AMCEs for different outcomes. An alternative is to use \code{\link[ggplot2]{facet_wrap}} for faceted graphics.}

\item{feature_headers}{A logical indicating whether to include headers for each feature to visually separate levels for each feature (beyond the color palette).}

\item{header_fmt}{A character string specifying a \code{fmt} argument to \code{\link[base]{sprintf}}, which will be used when generating the feature headers (if \code{feature_headers = TRUE}).}

\item{size}{A numeric value specifying point size in \code{\link[ggplot2]{geom_point}}.}

\item{xlab}{A label for the x-axis}

\item{ylab}{A label for the y-axis}

\item{legend_title}{A character string specifying a label for the legend.}

\item{xlim}{A two-element number vector specifying limits for the x-axis. If \code{NULL}, a default value is calculated from the data.}

\item{vline}{Optionally, a numeric value specifying an x-intercept for a vertical line. This can be useful in distinguishing the midpoint of the estimates (e.g., a zero line for AMCEs).}

\item{vline_color}{A character string specifying a color for the \code{vline}.}

\item{theme}{A ggplot2 theme object}

\item{\dots}{Ignored.}
}
\value{
A ggplot2 object
}
\description{
ggplot2-based plotting of conjoint AMCEs estimates and MMs
}
\details{
Two \code{plot} methods are implemented: one for \dQuote{amce} data frames of conjoint AMCE estimates (returned by \code{\link{amce}}), and another for \dQuote{mm} data frames as returned by \code{\link{mm}}. Both functions are basically identical but vary slightly in their superficial features (axis labels, etc.).
}
\examples{
\dontrun{
# load data
requireNamespace("ggplot2")
data("hainmueller")

# calculate MMs
d1 <- mm(hainmueller, ChosenImmigrant ~ Gender + Education + 
         LanguageSkills + CountryOfOrigin + Job + JobExperience + 
         JobPlans + ReasonForApplication + PriorEntry, id = ~ CaseID)

# plot MMs
plot(d1)

# MMs split by profile number
stacked <- do.call("rbind", lapply(1:3, function(x) {
    out <- mm(hainmueller[hainmueller$contest_no == x,], ChosenImmigrant ~ Gender + 
              Education + LanguageSkills + CountryOfOrigin + Job + JobExperience + 
              JobPlans + ReasonForApplication + PriorEntry, id = ~ CaseID)
    out$contest <- x
    out
}))
stacked$contest <- factor(stacked$contest, levels = 1:3, labels = c("First", "Second", "Third"))

## plot with grouping
plot(stacked, group = "contest", feature_headers = FALSE)

## plot with facetting
plot(stacked) + ggplot2::facet_wrap(~contest, nrow = 1L)

# estimate AMCEs
d2 <- cj(hainmueller, ChosenImmigrant ~ Gender + Education + 
         LanguageSkills + CountryOfOrigin + Job + JobExperience + 
         JobPlans + ReasonForApplication + PriorEntry, id = ~ CaseID)

# plot AMCEs
plot(d2)
}
}
\seealso{
\code{\link{amce}}, \code{\link{mm}}
}
