% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-cregg.R
\docType{package}
\name{cregg-package}
\alias{cregg-package}
\alias{cregg}
\alias{cj}
\title{Simple Conjoint Analyses and Visualization}
\usage{
cj(data, formula, id = NULL, weights = NULL, estimate = c("amce", "freqs",
  "mm", "differences"), by = NULL, ...)
}
\arguments{
\item{data}{A data frame containing variables specified in \code{formula}. All RHS variables should be factors; the base level for each will be used in estimation and for AMCEs the base level's AMCE will be zero.}

\item{formula}{A formula specifying a model to be estimated. All variables should be factors.}

\item{id}{An RHS formula specifying a variable holding respondent identifiers, to be used for clustering standard errors.}

\item{weights}{An (optional) RHS formula specifying a variable holding survey weights.}

\item{estimate}{A character string specifying an estimate type. Current options are average marginal component effects (or AMCEs, \dQuote{amce}, estimated via \code{\link{amce}}), display frequencies (\dQuote{freq}, estimated via \code{\link{freqs}}), marginal means (or AMMs, \dQuote{mm}, estimated via \code{\link{mm}}), or differences in AMCEs (\dQuote{differences}, via \code{\link{amce_diffs}}). Additional options may be made available in the future.}

\item{by}{A formula containing only RHS variables, specifying grouping factors over which to perform estimation.}

\item{\dots}{Additional arguments to \code{\link{amce}}, \code{\link{freqs}}, \code{\link{mm}}, or \code{\link{amce_diffs}}.}
}
\description{
Simple analyses of conjoint (factorial) experiments and visualization of results.
}
\details{
The main function \code{cj} is a convenience function wrapper around the underlying estimation functions that provide for average marginal component effects (AMCEs), by default, via the \code{\link{amce}} function, marginal means (MMs) via the \code{\link{mm}} function, and display frequencies via \code{\link{freqs}} and \code{\link{props}}. Additional estimands may be supported in the future through their own functions and through the \code{cj} interface. Plotting is provided via ggplot2 for both types of estimates.

The only additional functionality provided by \code{cj} over the underlying functions is the \code{by} argument, which will perform operations on subsets of \code{data}, returning a single data frame. This can be useful, for example, for evaluating profile spillover effects and subgroup results.
}
\examples{
\dontrun{
# load data
requireNamespace("ggplot2")
data("hainmueller")

# calculate MMs
d1 <- cj(hainmueller, ChosenImmigrant ~ Gender + Education + 
         LanguageSkills + CountryOfOrigin + Job + JobExperience + 
         JobPlans + ReasonForApplication + PriorEntry, id = ~ CaseID,
         estimate = "mm")

# plot MMs
plot(d1)

# MMs split by profile number
stacked <- mm(hainmueller[hainmueller$contest_no == x,], ChosenImmigrant ~ Gender + 
              Education + LanguageSkills + CountryOfOrigin + Job + JobExperience + 
              JobPlans + ReasonForApplication + PriorEntry, id = ~ CaseID, by = ~ contest_no)

## plot with grouping
plot(stacked, group = "contest_no", feature_headers = FALSE)

## plot with facetting
plot(stacked) + ggplot2::facet_wrap(~contest_no, nrow = 1L)

# estimate AMCEs
d2 <- cj(hainmueller, ChosenImmigrant ~ Gender + Education + 
         LanguageSkills + CountryOfOrigin + Job + JobExperience + 
         JobPlans + ReasonForApplication + PriorEntry, id = ~ CaseID)

# plot AMCEs
plot(d2)

# grouped operations
## examine profile ordering
x <- cj(hainmueller, ChosenImmigrant ~ Gender + Education + LanguageSkills,
        id = ~ CaseID, estimate = "mm", by = ~ contest_no)
plot(x, group = "contest_no", vline = 0.5)

## subgroup analysis
hainmueller$ethnosplit <- cut(hainmueller$ethnocentrism, 2)
x <- cj(na.omit(hainmueller), ChosenImmigrant ~ Gender + Education + LanguageSkills,
        id = ~ CaseID, estimate = "mm", by = ~ ethnosplit)
plot(x, group = "ethnosplit", vline = 0.5)
}
}
\seealso{
\code{\link{amce}} \code{\link{mm}} \code{\link{freqs}} \code{\link{plot.cj_amce}}
}
\author{
Thomas J. Leeper <thosjleeper@gmail.com>
}
\keyword{package}
