% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqs.R
\name{freqs}
\alias{freqs}
\alias{props}
\title{Conjoint feature frequencies}
\usage{
freqs(data, formula, id = NULL, weights = NULL, feature_order = NULL,
  feature_labels = NULL, level_order = c("ascending", "descending"), ...)

props(data, formula, id, weights = NULL, margin = NULL, ...)
}
\arguments{
\item{data}{A data frame containing variables specified in \code{formula}. All RHS variables should be factors.}

\item{formula}{An RHS formula specifying conjoint features to tabulate. All RHS variables should be factors.}

\item{id}{An RHS formula specifying a variable holding respondent identifiers, to be used for clustering standard errors.}

\item{weights}{An (optional) RHS formula specifying a variable holding survey weights.}

\item{feature_order}{An (optional) character vector specifying the names of feature (RHS) variables in the order they should be encoded in the resulting data frame.}

\item{feature_labels}{A named list of \dQuote{fancy} feature labels to be used in output. By default, the function looks for a \dQuote{label} attribute on each variable in \code{formula} and uses that for pretty printing. This argument overrides those attributes or otherwise provides fancy labels for this purpose. This should be a list with names equal to variables on the righthand side of \code{formula} and character string values; arguments passed here override variable attributes.}

\item{level_order}{A character string specifying levels (within each feature) should be ordered increasing or decreasing in the final output. This is mostly only consequential for plotting via \code{\link{plot.cj_mm}}, etc.}

\item{\dots}{Ignored.}

\item{margin}{A numeric value passed to \code{\link[base]{prop.table}}. If \code{NULL} overall proportions are calculated.}
}
\description{
Tabulate and visualize conjoint feature frequencies and proportions
}
\details{
These two functions provide slightly different functionality. \code{props} provides tidy proportion tables to examine cross-feature restrictions in conjoint designs that are not equally randomized. This enables, for example, tabulation and visualization of complete restrictions (where combinations of two or more features are not permitted), as well as calculation of AMCEs for constrained designs appropriately weighted by the display proportions for particular combinations of features.

\code{freqs} provides \emph{marginal} display frequencies, which are a descriptive check on the presentation of individual conjoint features (for example, to ensure equal or intentionally unequal appearance of levels). This is mostly useful for plotting functionality provided in \code{\link{plot.cj_freqs}}, which provides barcharts for the frequency with which each level of each feature was presented.
}
\examples{
data(immigration)
# display frequencies
freqs(immigration, ~ Gender + Education + LanguageSkills, id = ~ CaseID)

# restrictions
## check display proportions
props(immigration, ~ Job, id = ~ CaseID)
## check which combinations were not allowed
subset(props(immigration, ~ Job + Education, id = ~ CaseID), Proportion == 0)

}
\seealso{
\code{\link{plot.cj_mm}}
}
